Fix radix of parseInt
Places where we compare an int with a string the int will be converted to a string.
However when the value is assigned search params we have to parse the string as an int. Radix of parseInt defaults to 8, which seems to work. But just to be on the safe side we will explicitly use radix 10.

Auditors: jbr, mib, hwha