Rename pages
Otherwise they will clash with the forwarded api urls.

In the old merkur the endpoints don't start with /api. E.g. the endpoint in merkur-frontend called /api/conversions is called /conversions in old merkur. In order to affect the external customers as little as possible we want to maintain the old urls, but at the same time we can't avoid having /api in the url because of the way next.js works.

The solution is to have the proxy point the old api urls to the new api urls. But for that to work the pages have to be called something different from the api urls, otherwise they will never be hit.

Auditors: jbr, mib, hwha, akri